feat: Band connection and sync fix: Solved crash on Bio tab and on 3rd tab of header fix: Ring sync issue

+++ app/src/main/java/com/humotron/app/domain/modal/response/BioHackProgressResponse.kt
@@ -36,7 +36,7 @@ data class BioHackProgressResponse(
             @SerializedName("count")
             val count: Int?,
             @SerializedName("percentage")
-            val percentage: Int?
+            val percentage: Double?
         ) : Parcelable
 
         @Parcelize
